misc type changes
- don't use explicit 32 bit types (DWORD) for hook  functions and instead use size_t type
- make c_list_widget::update_list_items_from_mapping a const function

+++ xlive/Util/Hooks/Hook.cpp
@@ -84,13 +84,13 @@ void RetourClassFunc(BYTE *src, BYTE *restore, const unsigned int len)
 	);
 }
 
-void *VTableFunction(void *ClassPtr, DWORD index)
+void *VTableFunction(void *ClassPtr, size_t index)
 {
 	void **pVtable = *(void***)ClassPtr;
 	return pVtable[index];
 }
 
-void WriteBytes(DWORD destAddress, LPVOID bytesToWrite, const unsigned int numBytes)
+void WriteBytes(size_t destAddress, LPVOID bytesToWrite, const unsigned int numBytes)
 {
 	VirtualProtectAndExecutePatch((LPVOID)destAddress, numBytes, PAGE_EXECUTE_READWRITE,
 
@@ -98,12 +98,12 @@ void WriteBytes(DWORD destAddress, LPVOID bytesToWrite, const unsigned int numBy
 	);
 }
 
-void PatchCall(DWORD call_addr, DWORD new_function_ptr) {
-	DWORD callRelative = new_function_ptr - (call_addr + 5);
+void PatchCall(size_t call_addr, size_t new_function_ptr) {
+	size_t callRelative = new_function_ptr - (call_addr + 5);
 	WritePointer(call_addr + 1, reinterpret_cast<void*>(callRelative));
 }
 
-void NopFill(DWORD address, const unsigned int length)
+void NopFill(size_t address, const unsigned int length)
 {
 	VirtualProtectAndExecutePatch((LPVOID)address, length, PAGE_EXECUTE_READWRITE,
 
@@ -111,7 +111,7 @@ void NopFill(DWORD address, const unsigned int length)
 	);
 }
 
-void ReadBytesProtected(DWORD address, BYTE* buf, BYTE count)
+void ReadBytesProtected(size_t address, BYTE* buf, BYTE count)
 {
 	VirtualProtectAndExecutePatch((LPVOID)address, count, PAGE_EXECUTE_READWRITE,
 
@@ -119,11 +119,11 @@ void ReadBytesProtected(DWORD address, BYTE* buf, BYTE count)
 	);
 }
 
-void WritePointer(DWORD offset, const void *ptr) {
-	WriteValue<DWORD>(offset, *(DWORD*)&ptr);
+void WritePointer(size_t offset, const void *ptr) {
+	WriteValue<size_t>(offset, *(size_t*)&ptr);
 }
 
-void PatchWinAPICall(DWORD call_addr, DWORD new_function_ptr)
+void PatchWinAPICall(size_t call_addr, size_t new_function_ptr)
 {
 	BYTE call = 0xE8;
 	WriteValue(call_addr, call);
@@ -135,7 +135,7 @@ void PatchWinAPICall(DWORD call_addr, DWORD new_function_ptr)
 	WriteValue(call_addr + 5, padding);
 }
 
-VOID Codecave(DWORD destAddress, VOID(*func)(VOID), BYTE nopCount)
+VOID Codecave(size_t destAddress, VOID(*func)(VOID), BYTE nopCount)
 {
 	DWORD offset = (PtrToUlong(func) - destAddress) - 5;
